Clear input line when switching to vi mode
If vi_mode is set to 1 and the user leaves chat mode, clear the input
line.  Existing text on the input line cannot be sent as a message and
will also not be evaluated as a vi command in this situation, so leaving
the text on the input line would be confusing.